refactor(blockifier): move code from mod.rs to deprecated_syscall_executor
